+++ athithian_v/coding_task/doc-share/package.json
@@ -22,8 +22,11 @@
     "wrangler": "^3.60.3"
   },
   "dependencies": {
+    "@clerk/backend": "^1.13.2",
     "@clerk/clerk-sdk-node": "^5.0.41",
+    "@hono/clerk-auth": "^2.0.0",
     "@neondatabase/serverless": "^0.9.5",
+    "axios": "^1.7.7",
     "drizzle-orm": "^0.33.0",
     "hono": "^4.6.2"
   }
