+++ athithian_v/coding_task/doc-share/src/features/user/authentication.ts
@@ -1,11 +1,11 @@
 import { createClerkClient } from "@clerk/clerk-sdk-node";
 
-async function authenticate(signInToken:string, secretKey:string){
+async function authenticate(signInToken:string, secretKey:string, jwtKey:string){
     try {
         const clerk = createClerkClient({secretKey: secretKey});
-        const user = await clerk.clients.verifyClient(signInToken);
+        const user = await clerk.verifyToken(signInToken);
+        // await clerk.clients.verifyClient(signInToken);
         console.log(user);
-        
         return user;   
     } catch (error) {
         console.error(error);
